use boost stdint for portable integer types

+++ src/core/server/Resource.cc
@@ -143,7 +143,7 @@ Json Resource::Meta() const
 		if ( ec )
 			write_time = std::time(0) ;
 		
-		meta.Add( "last-modified", Json(static_cast<int>(write_time)) ) ;
+		meta.Add( "last-modified", Json(write_time) ) ;
 	}
 	
 	return meta ;
@@ -155,7 +155,7 @@ void Resource::SaveMeta(std::time_t modified) const
 	fs::create_directories( file.parent_path() ) ;
 	
 	Json meta = Meta() ;
-	meta.Add( "last-modified", Json(static_cast<int>(modified)) ) ;
+	meta.Add( "last-modified", Json(modified) ) ;
 
 	File out( file, 0600 ) ;
 	meta.Write( &out ) ;
